Converted root selection settings section to a class
